More work on intersections
Added code to work on the intersections. Trying to find out all the possible starting points for a person to build. TODO: Make a way for someone to put what intersection others are on and make they are out of the selection. If the player is on a tile only run the code for that tile and other possible matches.